+++ js/react/es5/DataFormsJS.js
@@ -1,5 +1,5 @@
 // @link https://www.dataformsjs.com
-// @version 5.13.1
+// @version 5.14.0
 // @author Conrad Sollitt (https://conradsollitt.com)
 // @license MIT
 (function () {